\name{setup}
\alias{setup}
\title{Project StartR}
\usage{
setup()
}
\description{
Prints 'This is a R package that seeds your R analyses with a consistent folder structure.
    This allows all projects to operate identically, making collaboration, maintenance, and sharing easier and more scalable.'.
}
\examples{
setup()
}
